requirements.txt
absl-py==2.1.0
altair==5.4.0
annotated-types==0.7.0
anyio==4.2.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
behave==1.2.6
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
bs4==0.0.2
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cssselect==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.29.2
distro==1.9.0
e==1.4.5
excel==1.0.1
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==24.3.25
fonttools==4.47.2
fqdn==1.5.1
fsspec==2024.6.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
jiter==0.5.0
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
loguru==0.7.2
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
narwhals==1.4.2
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.4
openai==1.40.8
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parameterized==0.9.0
parse==1.20.1
parse-type==0.6.2
parso==0.8.3
patsy==0.5.6
pdbp==1.5.0
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==17.0.0
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.17.2
pyk==0.3.2
pyktok==0.0.19
pynose==1.4.8
pyotp==2.9.0
pyparsing==3.1.1
pypdf==4.3.1
pyperclip==1.8.2
pyreadline3==3.4.1
PySocks==1.7.1
pytest==8.0.0
pytest-html==2.0.1
pytest-metadata==3.0.0
pytest-ordering==0.6
pytest-rerunfailures==13.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
redis==5.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.3
sbvirtualdisplay==1.3.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
selenium==4.17.2
seleniumbase==4.23.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.23.0
SQLAlchemy==2.0.32
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.37.1
streamlit-tags==1.2.8
sympy==1.13.1
tabcompleter==1.3.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
terminado==0.18.0
textblob==0.18.0.post0
tf_keras==2.16.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.42.4
trio==0.24.0
trio-websocket==0.11.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.18
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
WebSearcher==0.4.0
websocket-client==1.7.0
Werkzeug==3.0.2
widgetsnbextension==4.0.9
wrapt==1.16.0
wsproto==1.2.0
zipp==3.17.0
